+++ jwst/regtest/test_nirspec_mos_spec2.py
@@ -8,6 +8,7 @@ from jwst.barshadow import BarShadowStep
 import jwst.datamodels as dm
 from jwst.flatfield import FlatFieldStep
 from jwst.pathloss import PathLossStep
+from jwst.photom import PhotomStep
 from jwst.pipeline.collect_pipeline_cfgs import collect_pipeline_cfgs
 from jwst.stpipe import Step
 
@@ -210,3 +211,44 @@ def test_barshadow_source_type(jail, rtdata_module):
         if np.allclose(slit.barshadow, np.ones(slit.data.shape), equal_nan=True):
             bad_slits.append(idx)
     assert not bad_slits, f'Force to uniform failed for slits {bad_slits}'
+
+
+@pytest.mark.bigdata
+def test_photom_corrpars(jail, rtdata_module):
+    """Test for photom correction parameters"""
+    rtdata = rtdata_module
+    data = dm.open(rtdata.get_data('nirspec/mos/usf_wavecorr.fits'))
+
+    pls = PhotomStep()
+    corrected = pls.run(data)
+
+    pls.use_correction_pars = True
+    corrected_corrpars = pls.run(data)
+
+    bad_slits = []
+    for idx, slits in enumerate(zip(corrected.slits, corrected_corrpars.slits)):
+        corrected_slit, corrected_corrpars_slit = slits
+        if not np.allclose(corrected_slit.data, corrected_corrpars_slit.data, equal_nan=True):
+            bad_slits.append(idx)
+    assert not bad_slits, f'correction_pars failed for slits {bad_slits}'
+
+
+@pytest.mark.bigdata
+def test_photom_inverse(jail, rtdata_module):
+    """PhotomStep using inversion"""
+    rtdata = rtdata_module
+    data = dm.open(rtdata.get_data('nirspec/mos/usf_wavecorr.fits'))
+
+    pls = PhotomStep()
+    corrected = pls.run(data)
+
+    pls.inverse = True
+    corrected_inverse = pls.run(corrected)
+
+    bad_slits = []
+    for idx, slits in enumerate(zip(data.slits, corrected_inverse.slits)):
+        data_slit, corrected_inverse_slit = slits
+        non_nan = ~np.isnan(corrected_inverse_slit.data)
+        if not np.allclose(data_slit.data[non_nan], corrected_inverse_slit.data[non_nan]):
+            bad_slits.append(idx)
+    assert not bad_slits, f'Inversion failed for slits {bad_slits}'